Added spawning functionality in unopened containers based on conditions
Currently only events are used, as those are still being developed. Once at least 4 drawers, a bookcase door, and the cabinet are opened, the object will be spawned at a random unopened location. The other case is that the player has opened 4 drawers and the bookcase, but not opened the cabinet yet, in which case the object will always appear in the cabinet, making sure the player must familiarize themselves with each type of container.