Final day update
So, I can't test the sink handle clue. The Quest gets really unhappy when I try, but it doesn't seem to be the game itself (no crashing, functionality remains, etc., the Quest just loses it), so that's good. Aside from placing clutter in the room for the sake of fun little non-essential player interactions, the final day is done. Everything else works. Assuming the sink handle works as it should, we could build the final day and have a fully playable little game. If we want to build the first few days as a little tutorial, that's cool. I will say, though, that I'm not super keen on doing it on my own. I am very tired.