Changed GrabBase orientation and made key in the ice more visible
Changed HandBP's GrabBase so that instead of rotating about the palm of your hand, you now rotate about your wrist. Remember when people were talking about it hurting their wrists to turn knobs? Fixed that.

I also changed the material of the ice so that it's a bit easier to see the key inside it (and generally reduced the refraction—some people found it rather disorienting).