Sink/Clock Puzzle rework
Changed the rotations at which the sink handle will emit the ticking sounds and the times to which the clock hands need to be set to more easily-distinguishable values. Also reduced the wait-time necessary for the clock to consider the puzzle solved from 2.5 seconds to 1.5 seconds.
Added winding sound to play while the clock hands are moving back to their midnight position.

Reduced delay between game start and phone ringing from 15 seconds to 8 seconds.