Phone, Ice
Phone ringing is now independent of whether the cord has been broken (rotary phones only care if there's something on top of them... the ringing doesn't come from the receiver).

Water dripping from ice now spawns at the point on the ice closest to the ground (imagine water running down).
Key is now in the center of the ice instead of weirdly off-center.