Added a tray/tub for the ice to melt into
Brought in the BP_BasicCup into the scene, improved the way water fills the container. It is colored metallic silver firstly so that it stands out. It doesn't match anything else, hopefully it will strike a player as important as a result. Additionally, it being silver sends the message that it's made of metal and will conduct heat well, so lighting the plant on fire to melt the ice is an easier conclusion to come to.