Sink handle relocation. Light fixture material changes when it breaks
Sink handle has been moved to top-middle dresser drawer. 

Before, it was nearly impossible to see the shards of the light fixture after it shattered. Now, the transparent material is replaced with a white one after the fixture shatters (in the dimmer light, it looks more like the shards are reflecting light)